+++ app/src/main/java/eu/kanade/tachiyomi/ui/updates/UpdatesViewModel.kt
@@ -115,7 +115,13 @@ class UpdatesViewModel(
                 prefs.filterStarted,
                 prefs.filterBookmarked,
             )
-                .any { it != TriState.DISABLED }
+                .any { it != TriState.DISABLED } ||
+                prefs.filterExcludedScanlators ||
+                listOf(
+                    prefs.filterIncludedCategories,
+                    prefs.filterExcludedCategories,
+                )
+                    .any { it.isNotEmpty() }
         }
         .distinctUntilChanged()